requirements.txt
requests
apscheduler
python-dotenv
Flask
flask-ngrok
fastapi

pytest
pytest-asyncio
pytest-cov
moto[boto3]  
requests-mock